+++ src/aac_datasets/datasets/macs.py
@@ -1,25 +1,25 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
+import csv
 import logging
-import tqdm
 import os
 import os.path as osp
 import shutil
 import zipfile
 
-from dataclasses import asdict, astuple, dataclass, field, fields
+from dataclasses import dataclass, field, fields
 from functools import cached_property
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Any, Callable, Dict, Iterable, List, Optional, Union
 
 import torch
 import torchaudio
 import yaml
 
-from torch import nn, Tensor
+from torch import Tensor
+from torch.hub import download_url_to_file
 from torch.utils.data.dataset import Dataset
-from torchaudio.backend.common import AudioMetaData
-from torchaudio.datasets.utils import download_url
+from torchaudio.datasets.utils import validate_file
 
 
 logger = logging.getLogger(__name__)
@@ -27,19 +27,24 @@ logger = logging.getLogger(__name__)
 
 @dataclass
 class MACSItem:
+    # AAC attributes
     audio: Tensor = torch.empty((0,))
     captions: List[str] = field(default_factory=list)
-    tags: List[List[str]] = field(default_factory=list)
-    annotators_ids: List[str] = field(default_factory=list)
-    fname: Optional[str] = None
-    index: int = -1
     dataset: str = "macs"
+    fname: str = "unknown"
+    index: int = -1
+    subset: str = "unknown"
     sr: int = -1
+    # MACS-specific attributes
+    annotators_ids: List[str] = field(default_factory=list)
+    competences: List[float] = field(default_factory=list)
+    identifier: str = "unknwon"
+    scene_label: str = "unknown"
+    tags: List[List[str]] = field(default_factory=list)
 
 
 class MACS(Dataset):
-    r"""
-    Unofficial MACS pytorch dataset.
+    r"""Unofficial MACS pytorch dataset.
 
     .. code-block:: text
         :caption: Dataset folder tree
@@ -50,15 +55,20 @@ class MACS(Dataset):
             │    └── (3930 wav files, ~13GB)
             ├── LICENCE.txt
             ├── MACS.yaml
-            └── MACS_competence.csv
-
+            ├── MACS_competence.csv
+            └── tau_meta
+                ├── fold1_evaluate.csv
+                ├── fold1_test.csv
+                ├── fold1_train.csv
+                └── meta.csv
     """
 
-    AUDIO_MAX_LENGTH = 10.0  # in seconds
+    # Global
+    AUDIO_MAX_SIZE = 480001
+    AUDIO_MIN_SIZE = 479999
     AUDIO_N_CHANNELS = 2
     CLEAN_ARCHIVES: bool = False
     FORCE_PREPARE_DATA: bool = False
-    ITEM_TYPES = ("tuple", "dict", "dataclass")
     MAX_CAPTIONS_PER_AUDIO = {"full": 5}
     MIN_CAPTIONS_PER_AUDIO = {"full": 2}
     SAMPLE_RATE = 48000  # in Hz
@@ -69,10 +79,8 @@ class MACS(Dataset):
         root: str = ".",
         subset: str = "full",
         download: bool = False,
-        transforms: Optional[Dict[str, Optional[nn.Module]]] = None,
+        item_transform: Optional[Callable] = None,
         unfold: bool = False,
-        item_type: str = "tuple",
-        multihot_tags: bool = False,
         verbose: int = 0,
     ) -> None:
         """
@@ -84,17 +92,10 @@ class MACS(Dataset):
             defaults to "full".
         :param download: Download the dataset if download=True and if the dataset is not already downloaded.
             defaults to False.
-        :param transforms: The transform to apply to values.
-            Keys can be 'audio', 'captions' or 'tags'.
+        :param item_transform: The transform to apply to the global dict item. This transform is applied AFTER each field transform.
             defaults to None.
         :param unfold: If True, map captions to audio instead of audio to caption.
             defaults to True.
-        :param item_type: The type of the value returned by __getitem__.
-            Can be 'tuple', 'dict' or 'dataclass'.
-            defaults to 'tuple'.
-        :param multihot_tags: If True, tags are loaded as multihot tensors of 10 classes.
-            Otherwise the tags names are returns as a list of str.
-            defaults to False.
         :param verbose: Verbose level to use. Can be 0 or 1.
             defaults to 0.
         """
@@ -102,118 +103,118 @@ class MACS(Dataset):
             raise ValueError(
                 f"Invalid argument {subset=} for MACS. (expected one of {self.SUBSETS})"
             )
-        if item_type not in self.ITEM_TYPES:
-            raise ValueError(
-                f"Invalid argument {item_type=} for MACS. (expected one of {self.ITEM_TYPES})"
-            )
-
-        if transforms is None:
-            transforms = {}
-
-        accepted_keys = [field_.name for field_ in fields(MACSItem)]
-        for key in transforms.keys():
-            if key not in accepted_keys:
-                raise ValueError(
-                    f"Invalid transform {key=}. (expected one of {accepted_keys})"
-                )
 
         super().__init__()
-        self._root = root
-        self._subset = subset
-        self._download = download
-        self._transforms = transforms
-        self._unfold = unfold
-        self._item_type = item_type
-        self._multihot_tags = multihot_tags
-        self._verbose = verbose
+        self.__root = root
+        self.__subset = subset
+        self.__download = download
+        self.__item_transform = item_transform
+        self.__unfold = unfold
+        self.__verbose = verbose
 
-        self._all_infos = []
-        self._tag_to_idx = {}
-        self._idx_to_tag = {}
+        self.__annotator_id_to_competence = {}
+        self.__all_items = {}
+        self.__is_loaded = False
 
-        if self._download:
+        if self.__download:
             self._prepare_data()
         self._load_data()
 
-    def __len__(self) -> int:
-        return len(self._all_infos)
+    def get_field(self, key: str, index: Union[int, slice, Iterable[int]]) -> Any:
+        """Get a specific data field.
 
-    def __getitem__(self, index: int) -> Union[Tuple, Dict, MACSItem]:
-        kwargs = {field.name: self.get(field.name, index) for field in fields(MACSItem)}
-        item = MACSItem(**kwargs)
-
-        if self._item_type == "tuple":
-            return astuple(item)
-        elif self._item_type == "dict":
-            return asdict(item)
-        elif self._item_type == "dataclass":
-            return item
-        else:
-            raise ValueError(
-                f"Invalid item_type={self._item_type} for MACS. (expected one of {self.ITEM_TYPES})"
-            )
-
-    def get_raw(self, name: str, index: int) -> Any:
-        """Get a specific data field. (before transform)
-
-        :param name: The name of the value.
-            Can be 'audio', 'fpath', 'fname', 'captions', 'tags', 'annotators_ids'.
-        :param index: The index of the value in range [0, len(dataset)[.
+        :param key: The name of the field. Can be any attribute name of :class:`~MACSItem`.
+        :param index: The index or slice of the value in range [0, len(dataset)-1].
+        :returns: The field value. The type depends of the transform applied to the field.
         """
-        if not (0 <= index < len(self)):
-            raise IndexError(
-                f"Invalid argument {index=} for {self} (expected in range [0, {len(self)-1}])"
-            )
+        if isinstance(index, (int, slice)) and key in self.__all_items.keys():
+            return self.__all_items[key][index]
 
-        if name == "audio":
-            fpath = self.get("fpath", index)
-            value, sr = torchaudio.load(fpath)  # type: ignore
+        if isinstance(index, slice):
+            index = range(len(self))[index]
 
-            if value.nelement() == 0:
-                raise RuntimeError(f'Invalid audio number of elements in "{fpath}".')
+        if isinstance(index, Iterable):
+            return [self.get_field(key, idx) for idx in index]
+
+        if key == "audio":
+            fpath = self.get_field("fpath", index)
+            audio, sr = torchaudio.load(fpath)  # type: ignore
+
+            # Sanity check
+            if audio.nelement() == 0:
+                raise RuntimeError(
+                    f"Invalid audio number of elements in {fpath}. (expected {audio.nelement()=} > 0)"
+                )
             if sr != self.SAMPLE_RATE:
-                raise RuntimeError(f'Invalid sample rate in "{fpath}".')
+                raise RuntimeError(
+                    f"Invalid sample rate in {fpath}. (expected {self.SAMPLE_RATE} but found {sr=})"
+                )
+            return audio
 
-        elif name == "fpath":
-            fname = self.get("fname", index)
-            value = osp.join(self._dpath_audio, fname)
+        elif key == "audio_metadata":
+            fpath = self.get_field("fpath", index)
+            audio_metadata = torchaudio.info(fpath)  # type: ignore
+            return audio_metadata
 
-        elif name == "dataset":
-            value = "macs"
+        elif key == "competences":
+            annotators_ids = self.get_field("annotators_ids", index)
+            competences = [self.get_competence(id_) for id_ in annotators_ids]
+            return competences
 
-        elif name == "index":
-            value = index
+        elif key == "dataset":
+            return "macs"
 
-        elif name == "sr":
-            fpath = self.get("fpath", index)
-            audio_info: AudioMetaData = torchaudio.info(fpath)  # type: ignore
-            return audio_info.sample_rate
+        elif key == "fpath":
+            fname = self.get_field("fname", index)
+            fpath = osp.join(self._dpath_audio, fname)
+            return fpath
 
-        elif name in self._all_infos[index].keys():
-            value = self._all_infos[index][name]
+        elif key == "index":
+            return index
+
+        elif key == "num_channels":
+            audio_metadata = self.get_field("audio_metadata", index)
+            return audio_metadata.num_channels
+
+        elif key == "num_frames":
+            audio_metadata = self.get_field("audio_metadata", index)
+            return audio_metadata.num_frames
+
+        elif key == "sr":
+            audio_metadata = self.get_field("audio_metadata", index)
+            return audio_metadata.sample_rate
+
+        elif key == "subset":
+            return self.__subset
 
         else:
-            raise ValueError(f"Invalid value {name=} at {index=}.")
-        return value
+            keys = [field.name for field in fields(MACSItem)]
+            raise ValueError(
+                f"Invalid argument {key=} at {index=}. (expected one of {tuple(keys)})"
+            )
 
-    def get(self, name: str, index: int) -> Any:
-        """Get a specific data field. (after transform)
+    def get_competence(self, annotator_id: int) -> float:
+        """Get competence value for a specific annotator id."""
+        return self.__annotator_id_to_competence[annotator_id]
 
-        :param name: The name of the value.
-                Can be 'audio', 'fpath', 'fname', 'captions', 'tags', 'annotators_ids'.
-        :param index: The index of the value in range [0, len(dataset)[.
-        """
-        value = self.get_raw(name, index)
-        transform = self._transforms.get(name, None)
-        if transform is not None:
-            value = transform(value)
-        return value
+    @cached_property
+    def _dpath_archives(self) -> str:
+        return osp.join(self._dpath_data, "archives")
 
-    def tag_to_idx(self, tag: str) -> int:
-        return self._tag_to_idx[tag]
+    @cached_property
+    def _dpath_audio(self) -> str:
+        return osp.join(self._dpath_data, "audio")
 
-    def idx_to_tag(self, idx: int) -> str:
-        return self._idx_to_tag[idx]
+    @cached_property
+    def _dpath_data(self) -> str:
+        return osp.join(self.__root, "MACS")
+
+    @cached_property
+    def _dpath_tau_meta(self) -> str:
+        return osp.join(self._dpath_data, "tau_meta")
+
+    def _is_loaded(self) -> bool:
+        return self.__is_loaded
 
     def _is_prepared(self) -> bool:
         if not osp.isdir(self._dpath_audio):
@@ -228,77 +229,193 @@ class MACS(Dataset):
         fnames = os.listdir(self._dpath_audio)
         return len(data) == len(fnames)
 
+    def _load_data(self) -> None:
+        if not self._is_prepared():
+            raise RuntimeError(
+                f"Cannot load data: macs is not prepared in data root={self.__root}. Please use download=True in dataset constructor."
+            )
+
+        # Read data files
+        captions_fname = MACS_FILES["captions"]["fname"]
+        captions_fpath = osp.join(self._dpath_data, captions_fname)
+        if self.__verbose >= 2:
+            logger.debug(f"Reading captions file {captions_fname}...")
+
+        with open(captions_fpath, "r") as file:
+            caps_data = yaml.safe_load(file)
+
+        tau_meta_fname = "meta.csv"
+        tau_meta_fpath = osp.join(self._dpath_tau_meta, tau_meta_fname)
+        if self.__verbose >= 2:
+            logger.debug(
+                f"Reading Tau Urban acoustic scene meta file {tau_meta_fname}..."
+            )
+
+        with open(tau_meta_fpath, "r") as file:
+            reader = csv.DictReader(file, delimiter="\t")
+            tau_tags_data = list(reader)
+
+        competence_fname = "MACS_competence.csv"
+        competence_fpath = osp.join(self._dpath_data, competence_fname)
+        if self.__verbose >= 2:
+            logger.debug(f"Reading file {competence_fname}...")
+
+        with open(competence_fpath, "r") as file:
+            reader = csv.DictReader(file, delimiter="\t")
+            competences_data = list(reader)
+
+        # Store MACS data
+        all_items: Dict[str, List[Any]] = {
+            "fname": [item["filename"] for item in caps_data["files"]],
+            "captions": [
+                [subitem["sentence"] for subitem in item["annotations"]]
+                for item in caps_data["files"]
+            ],
+            "tags": [
+                [subitem["tags"] for subitem in item["annotations"]]
+                for item in caps_data["files"]
+            ],
+            "annotators_ids": [
+                [subitem["annotator_id"] for subitem in item["annotations"]]
+                for item in caps_data["files"]
+            ],
+        }
+        dataset_size = len(all_items["fname"])
+
+        # Build global mappings
+        fname_to_idx = {fname: i for i, fname in enumerate(all_items["fname"])}
+        annotator_id_to_competence = {
+            int(annotator["annotator_id"]): float(annotator["competence"])
+            for annotator in competences_data
+        }
+
+        # Store TAU Urban acoustic scenes data
+        tau_additional_keys = ("scene_label", "identifier")
+        all_items |= {
+            key: [None for _ in range(dataset_size)] for key in tau_additional_keys
+        }
+        tau_meta_fpath = osp.join(self._dpath_tau_meta, "meta.csv")
+        for tau_tags in tau_tags_data:
+            fname = osp.basename(tau_tags["filename"])
+            if fname in fname_to_idx:
+                idx = fname_to_idx[fname]
+                for key in tau_additional_keys:
+                    all_items[key][idx] = tau_tags[key]
+
+        if self.__unfold and self.MIN_CAPTIONS_PER_AUDIO[self.__subset] > 1:
+            all_infos_unfolded = {key: [] for key in all_items.keys()}
+
+            for i, captions in enumerate(all_items["captions"]):
+                for caption in captions:
+                    for key in all_items.keys():
+                        all_infos_unfolded[key].append(all_items[key][i])
+                    all_infos_unfolded["captions"] = [caption]
+
+            all_items = all_infos_unfolded
+
+        # Sanity checks
+        assert all(
+            all(value is not None for value in all_items[key])
+            for key in tau_additional_keys
+        )
+        assert all(len(values) == dataset_size for values in all_items.values())
+
+        # Set attributes
+        self.__all_items = all_items
+        self.__annotator_id_to_competence = annotator_id_to_competence
+        self.__is_loaded = True
+
+        if self.__verbose >= 1:
+            logger.info(f"{repr(self)} has been loaded. (len={len(self)})")
+
     def _prepare_data(self) -> None:
-        if not osp.isdir(self._root):
-            raise RuntimeError(f"Cannot find root directory '{self._root}'.")
+        if not osp.isdir(self.__root):
+            raise RuntimeError(f"Cannot find root directory '{self.__root}'.")
 
         os.makedirs(self._dpath_audio, exist_ok=True)
         os.makedirs(self._dpath_archives, exist_ok=True)
 
+        # Download MACS specific files
         for file_info in MACS_FILES.values():
             dpath = self._dpath_data
             fname = file_info["fname"]
             fpath = osp.join(dpath, fname)
 
             if not osp.isfile(fpath) or self.FORCE_PREPARE_DATA:
-                if self._verbose >= 1:
+                if self.__verbose >= 1:
                     logger.info(f"Downloading captions file '{fname}'...")
 
                 url = file_info["url"]
-                hash_value = file_info["hash"]
-                download_url(
+                download_url_to_file(
                     url,
-                    dpath,
-                    fname,
-                    hash_value,
-                    "md5",
-                    progress_bar=self._verbose >= 1,
-                    resume=True,
+                    fpath,
+                    progress=self.__verbose >= 1,
                 )
+
+            hash_value = file_info["hash"]
+            with open(fpath, "rb") as file:
+                valid = validate_file(file, hash_value, hash_type="md5")
+            if not valid:
+                raise RuntimeError(f"Invalid checksum for file {fname}.")
 
         captions_fpath = osp.join(self._dpath_data, MACS_FILES["captions"]["fname"])
         with open(captions_fpath, "r") as file:
             captions_data = yaml.safe_load(file)
         captions_data = captions_data["files"]
 
-        for i, infos in enumerate(MACS_ARCHIVES_FILES.values()):
+        # Download TAU Urban Sound audio archives files
+        for i, file_info in enumerate(MACS_ARCHIVES_FILES.values()):
             dpath = self._dpath_archives
-            zip_fname = infos["fname"]
+            zip_fname = file_info["fname"]
             zip_fpath = osp.join(dpath, zip_fname)
 
             if not osp.isfile(zip_fpath) or self.FORCE_PREPARE_DATA:
-                if self._verbose >= 1:
+                if self.__verbose >= 1:
                     logger.info(
                         f"Downloading audio zip file '{zip_fpath}'... ({i+1}/{len(MACS_ARCHIVES_FILES)})"
                     )
 
-                url = infos["url"]
-                hash_value = infos["hash"]
-                download_url(
+                url = file_info["url"]
+                download_url_to_file(
                     url,
-                    dpath,
-                    zip_fname,
-                    hash_value,
-                    "md5",
-                    progress_bar=self._verbose >= 1,
+                    zip_fpath,
+                    progress=self.__verbose >= 1,
                 )
 
+            hash_value = file_info["hash"]
+            with open(zip_fpath, "rb") as file:
+                valid = validate_file(file, hash_value, hash_type="md5")
+            if not valid:
+                raise RuntimeError(f"Invalid checksum for file {zip_fname}.")
+
+        # Extract files from TAU Urban Sound archives
         macs_fnames = dict.fromkeys(data["filename"] for data in captions_data)
-        for i, infos in enumerate(MACS_ARCHIVES_FILES.values()):
-            zip_fname = infos["fname"]
+        for i, (name, file_info) in enumerate(MACS_ARCHIVES_FILES.items()):
+            zip_fname = file_info["fname"]
             zip_fpath = osp.join(self._dpath_archives, zip_fname)
+
+            if self.__verbose >= 2:
+                logger.debug(
+                    f"Check to extract TAU Urban acoustic scenes archive {zip_fname=}..."
+                )
+
+            is_audio_archive = name.startswith("audio")
+            target_dpath = (
+                self._dpath_audio if is_audio_archive else self._dpath_tau_meta
+            )
 
             with zipfile.ZipFile(zip_fpath, "r") as file:
                 members_to_extract = [
                     member
                     for member in file.namelist()
-                    if osp.basename(member) in macs_fnames
-                    and not osp.isfile(
-                        osp.join(self._dpath_audio, osp.basename(member))
+                    # Extract member if file if in captions yaml file and if the audio file is not already downloaded
+                    if (
+                        (osp.basename(member) in macs_fnames or not is_audio_archive)
+                        and not osp.isfile(osp.join(target_dpath, osp.basename(member)))
                     )
                 ]
 
-                if self._verbose >= 1:
+                if self.__verbose >= 1:
                     logger.info(
                         f"Extracting {len(members_to_extract)}/{len(file.namelist())} audio files from ZIP file '{zip_fname}'... ({i+1}/{len(MACS_ARCHIVES_FILES)})"
                     )
@@ -307,11 +424,11 @@ class MACS(Dataset):
                     file.extractall(self._dpath_archives, members_to_extract)
                     for member in members_to_extract:
                         extracted_fpath = osp.join(self._dpath_archives, member)
-                        target_fpath = osp.join(self._dpath_audio, osp.basename(member))
+                        target_fpath = osp.join(target_dpath, osp.basename(member))
                         shutil.move(extracted_fpath, target_fpath)
 
         if self.CLEAN_ARCHIVES:
-            if self._verbose >= 1:
+            if self.__verbose >= 1:
                 logger.info(f"Removing archives files in {self._dpath_archives}...")
             shutil.rmtree(self._dpath_archives, ignore_errors=True)
 
@@ -321,107 +438,23 @@ class MACS(Dataset):
         audio_fpaths = [osp.join(self._dpath_audio, name) for name in audio_fnames]
         assert len(audio_fpaths) == len(macs_fnames)
 
-        if self._verbose >= 1:
+        if self.__verbose >= 1:
             logger.info(
                 f"{len(audio_fpaths)} audio files has been prepared for MACS dataset."
             )
 
-    def _load_data(self) -> None:
-        if not self._is_prepared():
-            raise RuntimeError(
-                f"Cannot load data: macs is not prepared in data root={self._root}. Please use download=True in dataset constructor."
-            )
+    def __getitem__(self, index: Union[int, slice]) -> Dict[str, Any]:
+        keys = [field.name for field in fields(MACSItem)]
+        item = {key: self.get_field(key, index) for key in keys}
+        if self.__item_transform is not None:
+            item = self.__item_transform(item)
+        return item
 
-        captions_fpath = osp.join(self._dpath_data, MACS_FILES["captions"]["fname"])
-        if self._verbose >= 1:
-            logger.debug(f'Loading captions file "{captions_fpath}"...')
-
-        with open(captions_fpath, "r") as file:
-            data = yaml.safe_load(file)
-
-        self._all_infos = [
-            {
-                "fname": item["filename"],
-                "captions": [subitem["sentence"] for subitem in item["annotations"]],
-                "tags": [subitem["tags"] for subitem in item["annotations"]],
-                "annotators_ids": [
-                    subitem["annotator_id"] for subitem in item["annotations"]
-                ],
-            }
-            for item in data["files"]
-        ]
-
-        all_tags = [item["tags"] for item in self._all_infos]
-        all_tags_flat = (
-            tag
-            for tags in all_tags
-            for tags_by_annotator in tags
-            for tag in tags_by_annotator
-        )
-        all_tags_uniq = dict.fromkeys(all_tags_flat)
-        self._tag_to_idx = {tag: i for i, tag in enumerate(all_tags_uniq)}
-        self._idx_to_tag = {i: tag for i, tag in enumerate(all_tags_uniq)}
-
-        if self._multihot_tags:
-            n_items = len(all_tags)
-            max_annotators = max(map(len, all_tags))
-            max_tag_by_annotator = max(
-                map(
-                    len,
-                    (
-                        tags_by_annotator
-                        for tags in all_tags
-                        for tags_by_annotator in tags
-                    ),
-                )
-            )
-            n_tags_uniq = len(all_tags_uniq)
-
-            multihots_tags_shape = (
-                n_items,
-                max_annotators,
-                max_tag_by_annotator,
-                n_tags_uniq,
-            )
-            multihots_tags = torch.zeros(multihots_tags_shape, dtype=torch.bool)
-
-            for i, tags in enumerate(tqdm.tqdm(all_tags, disable=self._verbose == 0)):
-                for j, tags_by_annotator in enumerate(tags):
-                    for k, tag in enumerate(tags_by_annotator):
-                        idx = self._tag_to_idx[tag]
-                        multihots_tags[i, j, k, idx] = True
-
-            for i, multihots in enumerate(multihots_tags):
-                self._all_infos[i]["tags"] = multihots
-
-        if self._unfold and self.MIN_CAPTIONS_PER_AUDIO["full"] > 1:
-            data_info_unfolded = []
-            for infos in self._all_infos:
-                captions = infos["captions"]
-                for caption in captions:
-                    new_infos = dict(infos)
-                    new_infos["captions"] = (caption,)
-                    data_info_unfolded.append(new_infos)
-
-            self._all_infos = data_info_unfolded
-
-        if self._verbose >= 1:
-            logger.info(f"{self.__class__.__name__} has been loaded. (len={len(self)})")
-
-    @cached_property
-    def _dpath_data(self) -> str:
-        return osp.join(self._root, "MACS")
-
-    @cached_property
-    def _dpath_audio(self) -> str:
-        return osp.join(self._dpath_data, "audio")
-
-    @cached_property
-    def _dpath_archives(self) -> str:
-        return osp.join(self._dpath_data, "archives")
+    def __len__(self) -> int:
+        return len(self.__all_items["captions"])
 
     def __repr__(self) -> str:
-        return "MACS()"
+        return f"MACS(size={len(self)})"
 
 
 # MACS-specific files links.
@@ -574,5 +607,6 @@ MACS_ARCHIVES_FILES = {
         "audio.2",
         "audio.3",
         "audio.9",
+        "meta",
     )
 }