+++ src/aac_datasets/utils.py
@@ -1,20 +1,22 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from typing import List, Tuple
+from typing import Any, Dict, List, Tuple
 
 from torch import Tensor
 from torch.nn.utils.rnn import pad_sequence
 
 
 class BasicCollate:
-    def __init__(self, pad_value: float = 0.0) -> None:
-        super().__init__()
-        self.pad_value = pad_value
+    """Collate object for :class:`~torch.utils.data.dataloader.DataLoader`."""
 
-    def __call__(self, batch: List[Tuple]) -> Tuple:
-        audio_batch = [item[0] for item in batch]
-        captions_batch = [item[1] for item in batch]
+    def __init__(self, audio_fill_value: float = 0.0) -> None:
+        super().__init__()
+        self.audio_fill_value = audio_fill_value
+
+    def __call__(self, batch: List[Dict[str, Any]]) -> Tuple:
+        audio_batch = [item["audio"] for item in batch]
+        captions_batch = [item["captions"] for item in batch]
 
         if len(audio_batch) == 0:
             raise ValueError("Cannot collate an empty list of audio.")
@@ -37,6 +39,6 @@ class BasicCollate:
         audio_batch = pad_sequence(
             audio_batch,
             batch_first=True,
-            padding_value=self.pad_value,
+            padding_value=self.audio_fill_value,
         )
         return audio_batch, captions_batch